+++ weblate/accounts/tests/test_notifications.py
@@ -22,7 +22,6 @@
 Tests for user handling.
 """
 
-from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User
 from django.core import mail
 
@@ -264,26 +263,3 @@ class NotificationTest(ViewTestCase):
             mail.outbox[0].subject,
             '[Weblate] Activity on your account at Weblate'
         )
-
-
-class RemoveAcccountTest(ViewTestCase):
-    def test_removal(self):
-        response = self.client.post(
-            reverse('remove')
-        )
-        self.assertRedirects(response, reverse('home'))
-        self.assertFalse(
-            User.objects.filter(username='testuser').exists()
-        )
-
-    def test_removal_change(self):
-        self.edit_unit(
-            'Hello, world!\n',
-            'Nazdar svete!\n'
-        )
-        # We should have some change to commit
-        self.assertTrue(self.subproject.repo_needs_commit())
-        # Remove account
-        self.test_removal()
-        # Changes should be committed
-        self.assertFalse(self.subproject.repo_needs_commit())